requirements.txt
pandas
numpy
nltk
scikit-learn
matplotlib
seaborn
xgboost
wordcloud
gensim
python-docx
pymupdf
streamlit
pdfminer.six
spacy
sentence-transformers
